add trigger type file on server
