+++ Composer/packages/server/src/models/asset/trigger.ts
@@ -1,0 +1,14 @@
+export const conceptLabels: { [key: string]: any } = {
+  'Microsoft.EventRule': {
+    title: 'Handle an Event',
+  },
+  'Microsoft.IntentRule': {
+    title: 'Handle an Intent',
+  },
+  'Microsoft.UnknownIntentRule': {
+    title: 'Handle Unknown Intent',
+  },
+  'Microsoft.ConversationUpdateActivityRule': {
+    title: 'Handle ConversationUpdate',
+  },
+};
